fix NumberFormatException due to Java being Java
A large negative number (32-bits with a leading 1) represented as
hex cannot be converted to an int directly with Integer.parseInt().
It must be converted to a long and then cast as an int. Why? Stop
asking silly questions.

+++ src/base/src/com/aerofs/base/id/OrganizationID.java
@@ -29,7 +29,9 @@ public class OrganizationID extends IntegerID
 
     public static OrganizationID fromHexString(String hexString)
     {
-        return new OrganizationID(Integer.parseInt(hexString, 16));
+        // N.B. parse as long and then cast to int, because large negative
+        // numbers will cause Integer.parseInt to throw a NumberFormatException
+        return new OrganizationID((int)Long.parseLong(hexString, 16));
     }
 
     public OrganizationID(int i)
